requirements.txt
altgraph==0.17.4
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
python-barcode==0.15.1
pywin32-ctypes==0.2.3
